Fix can't change airplane mode from GlobalActions.
A PhoneStateListener is registered in the constructor of
GlobalActions which, on rare occasions, leads to a callback being
received immediately. In these cases because mHasTelephony has not
been initialized, airplane mode is not refreshed in the callback
which disables the option to set airplane mode from GlobalActions.

To fix this issue mHasTelephony has to be set before registering
the listener.

Change-Id: If117a879d3244414b369a1711bf46526b7a5ae24

Conflicts:
	policy/src/com/android/internal/policy/impl/GlobalActions.java