framework/base: Do not kill the supplicant on a request to stop it
Supplicant is killed on obtaining the first TERMINATE event when
requested to stop from the framework. This shall result in improper
shutdown of the supplicant without proper cleanup of the resources
allocated for all the interfaces.
Moreover, there is no point in killing the supplicant when the
request itself is to to stop the same. Thus, do not kill the
supplicant on the TERMINATE event.

Change-Id: I954bb09faa94252eea53a14b090a6cff610c5964
CRs-Fixed: 601759

Conflicts:
	wifi/java/android/net/wifi/WifiStateMachine.java